+++ NEWS.md
@@ -1,7 +1,12 @@
+
+* Reworked package functions to support a single flagship function, get_reddit_content
+* Renamed underlying functions to be more specific
+* New unit tests for existing functionality
+
 # redditr 0.0.2
 
 * Added unit tests for interactions between url constructing and data importing
-* Added functionality for importaing data from constructed urls
+* Added functionality for importing data from constructed urls
 
 # redditr 0.0.1